pyproject.toml
[project]
name = "simply-mneme"
version = "0.1.0"
description = "Storage layer implementing noema's Unified Content Model (UCM)"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "aiofiles>=24.1",
]

[project.optional-dependencies]
nous = ["simply-nous"]

[tool.uv.sources]
simply-nous = { path = "../../nous", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mneme"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
